% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.array.R
\name{bind.array}
\alias{bind.array}
\title{bind.array: bind two 4d arrays to one 4d array.}
\usage{
bind.array(x, y)
}
\arguments{
\item{x}{a list obtained by the GoogleImage2array function.}

\item{y}{a list obtained by the GoogleImage2array function.}
}
\value{
array
}
\description{
This function is to bind two arrays to one array
along the first dimension,
which obtained by the GoogleImage2array function.
}
\examples{
\donttest{
library(EBImage)

# Simple examples
query <- "persian cat"
CatImg <- GoogleImage2array(query)

query <- "Shiba inu"
DogImg <- GoogleImage2array(query)

#bind arrays
Dat <- bind.array(CatImg, DogImg)
str(Dat)

}

}
\author{
Satoshi Kume
}
