% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.array.R
\name{display.array}
\alias{display.array}
\title{display.array: display 4d array as a tiled image.}
\usage{
display.array(x, Save = FALSE, file_path = NULL)
}
\arguments{
\item{x}{a list obtained by the GoogleImage2array function.}

\item{Save}{a logical. Whether to save images locally or not.
if TRUE, save locally.}

\item{file_path}{a character. a directory to save the image file.}
}
\value{
image
}
\description{
This function is to create the R array/tensor from
2D image obtained from Google image search.
}
\examples{
\donttest{
library(EBImage)

# Simple examples
query <- "persian cat"
CatImg <- GoogleImage2array(query)

#show images
display.array(CatImg)

query <- "Shiba inu"
DogImg <- GoogleImage2array(query)

#show images
display.array(DogImg)

}

}
\author{
Satoshi Kume
}
