% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.spiral.R
\name{display.spiral}
\alias{display.spiral}
\title{display.spiral: display 4d array as a spiral image.}
\usage{
display.spiral(x, Save = FALSE, SaveFormat = "png", file_path = NULL)
}
\arguments{
\item{x}{a list obtained by the GoogleImage2array function.}

\item{Save}{a logical. Whether to save images locally or not.
if TRUE, save locally.}

\item{SaveFormat}{a character of format; png or pdf.}

\item{file_path}{a character. a directory to save the image file.}
}
\value{
image
}
\description{
This function is to create a spiral imagethe
from R array/tensor obtained from Google image search.
}
\examples{
\donttest{
library(EBImage)

# Simple examples
query <- "persian cat"
CatImg <- GoogleImage2array(query)

#show images
display.spiral(CatImg)

}

}
\author{
Satoshi Kume
}
