% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleImage2array.world.R
\name{GoogleImage2array.world}
\alias{GoogleImage2array.world}
\title{GoogleImage2array.world}
\usage{
GoogleImage2array.world(Query, wh = 64, Col = TRUE, verbose = FALSE)
}
\arguments{
\item{Query}{a character vector to search images}

\item{wh}{a value of pixels in height and width.}

\item{Col}{a logical. Whether to handle color or gray images.
if TRUE, use color mode.}

\item{verbose}{Verbosity mode (FALSE: silent, TRUE: progress).}
}
\value{
array
}
\description{
This function is to gather images via 10 countries and
create the R array/tensor from 2D images obtained.
}
\examples{
\dontrun{
library(EBImage)

# Simple examples
query <- "persian cat"
CatImg <- GoogleImage2array.world(query)

#show info
str(CatImg)

}

}
\author{
Satoshi Kume
}
